backend/requirements.txt
boto3==1.34.11
botocore==1.34.11
dnspython==2.4.2
jmespath==1.0.1
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
